requirements.txt
aerich==0.6.3
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
asyncmy==0.2.9
certifi==2024.7.4
click==8.1.7
colorama==0.4.6
dictdiffer==0.9.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iso8601==1.1.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysql-connector-python==9.0.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pypika-tortoise==0.1.6
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
tomlkit==0.13.0
tortoise-orm==0.21.5
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uvicorn==0.30.3
watchfiles==0.22.0
websockets==12.0
